🛠️(refactor) 중복되는 UseCase, Repository 공유하는 방법으로 변경
