+++ roome/roome/Application/DIManager.swift
@@ -58,72 +58,50 @@ class DIManager {
         DIContainer.shared.register(WelcomeViewModel.self, dependency: welcomeViewModel)
         DIContainer.shared.register(WelcomeSignUPViewController.self, dependency: welcomeSignUPViewController)
         
-        let roomCountRepository = RoomCountRepository()
-        let roomCountUseCase = RoomCountUseCase(repository: roomCountRepository)
-        let roomCountViewModel = RoomCountViewModel(usecase: roomCountUseCase)
+        let profileSelectRepository = ProfileSelectRepository()
+        let profileSelectUseCase = ProfileSelectUseCase(repository: profileSelectRepository)
+        
+        let roomCountViewModel = RoomCountViewModel(usecase: profileSelectUseCase)
         let roomCountViewController = RoomCountViewController(viewModel: roomCountViewModel)
-        DIContainer.shared.register(RoomCountRepository.self, dependency: roomCountRepository)
-        DIContainer.shared.register(RoomCountUseCase.self, dependency: roomCountUseCase)
-        DIContainer.shared.register(RoomCountViewModel.self, dependency: roomCountViewModel)
         DIContainer.shared.register(RoomCountViewController.self, dependency: roomCountViewController)
         
-        let genreRepository = GenreRepository()
-        let genreUseCase = GenreUseCase(repository: genreRepository)
-        let genreViewModel = GenreViewModel(useCase: genreUseCase)
+        let genreViewModel = GenreViewModel(useCase: profileSelectUseCase)
         let genreViewController = GenreViewController(viewModel: genreViewModel)
         DIContainer.shared.register(GenreViewController.self, dependency: genreViewController)
         
-        let mbtiRepository = MbtiRepository()
-        let mbtiUseCase = MbtiUseCase(repository: mbtiRepository)
-        let mbtiViewModel = MBTIViewModel(useCase: mbtiUseCase)
+        let mbtiViewModel = MBTIViewModel(useCase: profileSelectUseCase)
         let mbtiViewController = MBTIViewController(viewModel: mbtiViewModel)
         DIContainer.shared.register(MBTIViewController.self, dependency: mbtiViewController)
         
-        let strengthRepository = StrengthRepository()
-        let strengthUseCase = StrengthUseCase(repository: strengthRepository)
-        let strengthViewModel = StrengthViewModel(useCase: strengthUseCase)
+        let strengthViewModel = StrengthViewModel(useCase: profileSelectUseCase)
         let strengthViewController = StrengthViewController(viewModel: strengthViewModel)
         DIContainer.shared.register(StrengthViewController.self, dependency: strengthViewController)
         
-        let themeRepository = ThemeSelectRepository()
-        let themeUseCase = ThemeSelectUseCase(repository: themeRepository)
-        let themeViewModel = ImportantFactorViewModel(useCase: themeUseCase)
+        let themeViewModel = ImportantFactorViewModel(useCase: profileSelectUseCase)
         let themeSelectViewController = ImportantFactorViewController(viewModel: themeViewModel)
         DIContainer.shared.register(ImportantFactorViewController.self, dependency: themeSelectViewController)
         
-        let horrorRepository = HorrorThemeRepository()
-        let horrorUseCase = HorrorThemeUseCase(repository: horrorRepository)
-        let horrorViewModel = HorrorPositionViewModel(useCase: horrorUseCase)
+        let horrorViewModel = HorrorPositionViewModel(useCase: profileSelectUseCase)
         let horrorPositionViewController = HorrorPositionViewController(viewModel: horrorViewModel)
         DIContainer.shared.register(HorrorPositionViewController.self, dependency: horrorPositionViewController)
         
-        let hintRepository = HintRepository()
-        let hintUseCase = HintUseCase(repository: hintRepository)
-        let hintViewModel = HintViewModel(useCase: hintUseCase)
+        let hintViewModel = HintViewModel(useCase: profileSelectUseCase)
         let hintViewController = HintViewController(viewModel: hintViewModel)
         DIContainer.shared.register(HintViewController.self, dependency: hintViewController)
         
-        let deviceLockRepository = DeviceLockRepository()
-        let deviceLockUseCase = DeviceLockUseCase(repository: deviceLockRepository)
-        let deviceLockViewModel = DeviceAndLockViewModel(useCase: deviceLockUseCase)
+        let deviceLockViewModel = DeviceAndLockViewModel(useCase: profileSelectUseCase)
         let deviceAndLockViewController = DeviceAndLockViewController(viewModel: deviceLockViewModel)
         DIContainer.shared.register(DeviceAndLockViewController.self, dependency: deviceAndLockViewController)
         
-        let activityRepository = ActivityRepository()
-        let activityUseCase = ActivityUseCase(repository: activityRepository)
-        let activityViewModel = ActivityViewModel(useCase: activityUseCase)
+        let activityViewModel = ActivityViewModel(useCase: profileSelectUseCase)
         let activityViewController = ActivityViewController(viewModel: activityViewModel)
         DIContainer.shared.register(ActivityViewController.self, dependency: activityViewController)
         
-        let dislikeRepository = DislikeRepository()
-        let dislikeUseCase = DislikeUseCase(repository: dislikeRepository)
-        let dislikeViewModel = DislikeViewModel(useCase: dislikeUseCase)
+        let dislikeViewModel = DislikeViewModel(useCase: profileSelectUseCase)
         let dislikeViewController = DislikeViewController(viewModel: dislikeViewModel)
         DIContainer.shared.register(DislikeViewController.self, dependency: dislikeViewController)
         
-        let colorRepository = ColorRepository()
-        let colorUseCase = ColorUseCase(repository: colorRepository)
-        let colorViewModel = ColorSelectViewModel(useCase: colorUseCase)
+        let colorViewModel = ColorSelectViewModel(useCase: profileSelectUseCase)
         let colorSelectViewController = ColorSelectViewController(viewModel: colorViewModel)
         DIContainer.shared.register(ColorSelectViewController.self, dependency: colorSelectViewController)
