--- roome/roome/ProfileSelect/Domain/Repository/ColorRepositoryType.swift
@@ -1,12 +1,0 @@
-//
-//  ColorRepositoryType.swift
-//  roome
-//
-//  Created by minsong kim on 5/31/24.
-//
-
-import Foundation
-
-protocol ColorRepositoryType {
-    func registerColor(id: Int) async throws
-}
