--- roome/roome/ProfileSelect/Domain/Repository/DislikeRepositoryType.swift
@@ -1,12 +1,0 @@
-//
-//  DislikeRepositoryType.swift
-//  roome
-//
-//  Created by minsong kim on 5/30/24.
-//
-
-import Foundation
-
-protocol DislikeRepositoryType {
-    func registerDislikes(ids: [Int]) async throws
-}
