--- roome/roome/ProfileSelect/Domain/Repository/GenreRepositoryType.swift
@@ -1,12 +1,0 @@
-//
-//  GenreRepositoryType.swift
-//  roome
-//
-//  Created by minsong kim on 5/30/24.
-//
-
-import Foundation
-
-protocol GenreRepositoryType {
-    func registerGenre(ids: [Int]) async throws
-}
