--- roome/roome/ProfileSelect/Domain/Repository/HintRepositoryType.swift
@@ -1,12 +1,0 @@
-//
-//  HintRepositoryType.swift
-//  roome
-//
-//  Created by minsong kim on 5/30/24.
-//
-
-import Foundation
-
-protocol HintRepositoryType {
-    func registerHintType(id: Int) async throws
-}
