--- roome/roome/ProfileSelect/Domain/Repository/HorrorThemeRepositoryType.swift
@@ -1,12 +1,0 @@
-//
-//  HorrorThemeRepositoryType.swift
-//  roome
-//
-//  Created by minsong kim on 5/30/24.
-//
-
-import Foundation
-
-protocol HorrorThemeRepositoryType {
-    func registerHorrors(id: Int) async throws
-}
