--- roome/roome/ProfileSelect/Domain/Repository/StrengthRepositoryType.swift
@@ -1,12 +1,0 @@
-//
-//  StrengthRepositoryType.swift
-//  roome
-//
-//  Created by minsong kim on 5/30/24.
-//
-
-import Foundation
-
-protocol StrengthRepositoryType {
-    func registerStrengths(ids: [Int]) async throws
-}
