--- roome/roome/ProfileSelect/Domain/Repository/ThemeSelectRepositoryType.swift
@@ -1,12 +1,0 @@
-//
-//  ThemeSelectRepositoryType.swift
-//  roome
-//
-//  Created by minsong kim on 5/30/24.
-//
-
-import Foundation
-
-protocol ThemeSelectRepositoryType {
-    func registerThemes(ids: [Int]) async throws
-}
