--- roome/roome/ProfileSelect/Domain/UseCase/HintUseCase.swift
@@ -1,20 +1,0 @@
-//
-//  HintUseCase.swift
-//  roome
-//
-//  Created by minsong kim on 5/30/24.
-//
-
-import Foundation
-
-class HintUseCase {
-    private let repository: HintRepositoryType
-    
-    init(repository: HintRepositoryType) {
-        self.repository = repository
-    }
-    
-    func hintWithAPI(id: Int) async throws {
-        try await repository.registerHintType(id: id)
-    }
-}
