+++ roome/roome/ProfileSelect/Presentation/ViewModel/MBTIViewModel.swift
@@ -31,9 +31,9 @@ class MBTIViewModel {
     private var withoutButtonState = false
     private var list: [Int: Int] = [0: -1, 1: -1, 2: -1, 3: -1]
     private var count: Int = 0
-    private var useCase: MbtiUseCase
+    private var useCase: ProfileSelectUseCaseType
     
-    init(useCase: MbtiUseCase) {
+    init(useCase: ProfileSelectUseCaseType) {
         self.useCase = useCase
         self.input = Input()
         self.output = Output()
